requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.11
dlib==19.24.6
duckduckgo_search==8.0.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
es_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.8.0/es_core_news_sm-3.8.0-py3-none-any.whl#sha256=e451a83d6df79b87e9eed0cb553f03e99e36a3bab18a7b79f0dcfd1fdf875e12
es_dep_news_trf @ https://github.com/explosion/spacy-models/releases/download/es_dep_news_trf-3.8.0/es_dep_news_trf-3.8.0-py3-none-any.whl#sha256=afe76019ca75827db18f14f282f4678b1544277b29dcba8011b400837b19dcb4
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.18.0
fsspec==2025.3.2
fuzzywuzzy==0.18.0
google-api-core==2.24.2
google-api-python-client==2.48.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
Google-Images-Search==1.4.7
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
imutils==0.5.4
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
Levenshtein==0.27.1
lxml==5.3.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pefile==2023.2.7
pillow==11.2.1
preshed==3.0.9
primp==0.14.0
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
pyfiglet==0.8.post1
Pygments==2.19.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.2
pyparsing==3.2.3
python-Levenshtein==0.27.1
python-resize-image==1.1.20
pywin32-ctypes==0.2.3
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rsa==4.9
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
spacy==3.8.5
spacy-curated-transformers==0.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.13.1
termcolor==1.1.0
thinc==8.3.6
torch==2.6.0
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
